refactor(Bipan): :rotating_light: remove log sighandler + fix signedness and extension warnings with explicit casts

+++ src/jni/spoofer.hpp
@@ -5,8 +5,8 @@
 
 int uname_spoofer(struct utsname* buf);
 int create_spoofed_file(const char* fake_content);
-long clean_proc_maps(int dirfd, const char* pathname, int flags, mode_t mode);
-long clean_proc_smaps(int dirfd, const char* pathname, int flags, mode_t mode);
-long clean_proc_mounts(int dirfd, const char* pathname, int flags, mode_t mode);
+int clean_proc_maps(int dirfd, const char* pathname, int flags, mode_t mode);
+int clean_proc_smaps(int dirfd, const char* pathname, int flags, mode_t mode);
+int clean_proc_mounts(int dirfd, const char* pathname, int flags, mode_t mode);
 
 #endif
